[add] 用户管理功能  1、查询用户 2、删除用户 添加管理员全局校验

+++ src/main/java/com/mingligu/controller/UserController.java
@@ -1,16 +1,17 @@
 package com.mingligu.controller;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.mingligu.constant.UserContact;
 import com.mingligu.model.User;
 import com.mingligu.model.request.UserLoginRequest;
 import com.mingligu.model.request.UserRegisterRequest;
 import com.mingligu.service.UserService;
 import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -77,7 +78,11 @@ public class UserController {
      * @return 返回用户数据集合
      */
     @GetMapping("/search")
-    public List<User> searchUsers(String userName) {
+    public List<User> searchUsers(String userName, HttpServletRequest request) {
+        if (isDefault(request)) {
+            return new ArrayList<>();
+        }
+
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
         if (StringUtils.isNotBlank(userName)) {
             queryWrapper.like("userName", userName);
@@ -92,12 +97,40 @@ public class UserController {
      * @return true 删除用户 / false 删除失败
      */
     @PostMapping("/delete")
-    public boolean deleteUser(@RequestBody long id) {
+    public boolean deleteUser(long id, HttpServletRequest request) {
+        if (isDefault(request)) {
+            return false;
+        }
+
         if (id <= 0) {
             return false;
         }
 
         return userService.removeById(id);
+    }
+
+    /**
+     * 是否是普通用户
+     * @param request Http请求对象
+     * @return 默认成员 true / 管理员 false
+     */
+    private boolean isDefault(HttpServletRequest request) {
+        return !isAdmin(request);
+    }
+
+    /**
+     * 是否是管理员
+     * @param request Http请求对象
+     * @return  管理员 true / 默认成员 false
+     */
+    private boolean isAdmin(HttpServletRequest request) {
+        //仅超级管理员用户可以查询
+        Object objUser = request.getSession().getAttribute(UserContact.USER_LOGIN_STATE);
+        User user = (User) objUser;
+        if (null == user || user.getUserRole() != UserContact.ADMIN_ROLE) {
+            return false;
+        }
+        return true;
     }
 
 }
